requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-factory-boy==1.0.0
django-filter==22.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==14.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
